chore: moved all css to styles folder
